requirements.txt
requests
python-binance
regex
async_timeout
aiohttp
dateparser
multidict
